Refactor: Consolidate login use cases
