--- domain/src/main/java/com/london/domain/usecase/LoggedInUseCase.kt
@@ -1,10 +1,0 @@
-package com.london.domain.usecase
-
-import com.london.domain.repository.AuthRepository
-import javax.inject.Inject
-
-class LoggedInUseCase @Inject constructor(
-    private val authRepository: AuthRepository
-) {
-    suspend fun invoke() = authRepository.isLoggedIn()
-}
