+++ domain/src/main/java/com/london/domain/usecase/login/AuthenticationUseCase.kt
@@ -1,0 +1,17 @@
+package com.london.domain.usecase.login
+
+import com.london.domain.repository.AuthRepository
+import javax.inject.Inject
+
+class AuthenticationUseCase @Inject constructor(
+    private val authRepository: AuthRepository
+) {
+    suspend fun loginAsGuest() = authRepository.loginAsGuest()
+
+    suspend fun login(username: String, password: String) =
+        authRepository.login(username, password)
+
+    suspend fun logout() = authRepository.logout()
+
+    suspend fun isLoggedIn() = authRepository.isLoggedIn()
+}
